Move prefab loaders to components modules
